+++ src/utils.ts
@@ -28,34 +28,14 @@ export default class Utils {
     return {
       darkContrast: workspaceConfiguration.get<string>("darkContrast"),
       lightContrast: workspaceConfiguration.get<string>("lightContrast"),
-      darkWorkbench: workspaceConfiguration.get<string>("darkWorkbench"),
-      lightWorkbench: workspaceConfiguration.get<string>("lightWorkbench"),
-      darkSelection: workspaceConfiguration.get<string>("darkSelection"),
-      lightSelection: workspaceConfiguration.get<string>("lightSelection"),
-      darkCursor: workspaceConfiguration.get<string>("darkCursor"),
-      lightCursor: workspaceConfiguration.get<string>("lightCursor"),
-      italicKeywords: workspaceConfiguration.get<boolean>("italicKeywords"),
-      italicComments: workspaceConfiguration.get<boolean>("italicComments"),
-      diagnosticTextBackgroundOpacity: workspaceConfiguration.get<string>(
-        "diagnosticTextBackgroundOpacity"
-      ),
       highContrast: workspaceConfiguration.get<boolean>("highContrast"),
     };
   } // }}}
   isDefaultConfiguration(configuration: Configuration): boolean {
     // {{{
     return (
-      configuration.italicKeywords === false &&
-      configuration.italicComments === true &&
-      configuration.lightWorkbench === "material" &&
-      configuration.darkWorkbench === "material" &&
       configuration.lightContrast === "medium" &&
       configuration.darkContrast === "medium" &&
-      configuration.darkCursor === "white" &&
-      configuration.lightCursor === "black" &&
-      configuration.darkSelection === "grey" &&
-      configuration.lightSelection === "grey" &&
-      configuration.diagnosticTextBackgroundOpacity === "0%" &&
       configuration.highContrast === false
     );
   } // }}}