+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to the "everforest" extension will be documented in this file.
 
+## Unreleased
+
+- Remove `everforest.darkSelection` and `everforest.lightSelection` customization options.
+- Always use the default selection colors in all workbench styles.
+- Remove `everforest.darkCursor` and `everforest.lightCursor` customization options.
+- Always use the default cursor colors in all workbench styles.
+- Remove `everforest.diagnosticTextBackgroundOpacity` customization option.
+- Always use the default diagnostic text background opacity in all workbench styles.
+- Remove `everforest.italicKeywords` and `everforest.italicComments` customization options.
+- Keywords and comments are never italicized.
+- Remove `everforest.darkWorkbench` and `everforest.lightWorkbench` customization options.
+- Always use the flat workbench style.
+
 ## 0.3.0
 
 - Change `activationEvents`.